remove bullet

+++ Pangu.spritebuilder/Source/ScriptLoader.m
@@ -13,7 +13,7 @@
 
 #pragma mark init
 +(ScriptLoader*) loaderOfLevel:(NSInteger)level{
-    NSString *filePath = [[NSBundle mainBundle] pathForResource:[NSString stringWithFormat:@"level%d", level] ofType:@"json"];
+    NSString *filePath = [[NSBundle mainBundle] pathForResource:[NSString stringWithFormat:@"level%ld", (long)level] ofType:@"json"];
     return [[ScriptLoader alloc] initWithFile:filePath];
 }
